+++ CHANGELOG.md
@@ -5,11 +5,12 @@
 * Use Reduct's colours for UI, [PR-7](https://github.com/reduct-storage/web-console/pull/7)
 * Implement changing bucket settings, [PR-9](https://github.com/reduct-storage/web-console/pull/9)
 * Show error in CreateOrUpdate form if don't get settings, [PR-10](https://github.com/reduct-storage/web-console/pull/10)
-* Add Login Form for authentication with token, [PR0-14](https://github.com/reduct-storage/web-console/pull/14)
+* Add Login Form for authentication with token, [PR0-13](https://github.com/reduct-storage/web-console/pull/13)
 
 **Bugs**:
 
 * Fix bigint conversions in bucket settings, [PR-11](https://github.com/reduct-storage/web-console/pull/11)
+* Use SI sizes for bucket settings, [PR0-13](https://github.com/reduct-storage/web-console/pull/13) 
 
 **Other**: